pyproject.toml
[project]
name = "bbrain.iac"
version = "0.1.0"
description = ""
authors = [
    {name = "Jonathan Scherrer", email = "jonathan.s.scherrer@gmail.com"},
]
dependencies = [
    "ovh>=1.0.0",
    "ruamel.yaml>=0.17.21",
    "requests>=2.28.1",
    "flatten-dict>=0.4.2",
    "deepdiff>=5.8.1",
    "click>=8.1.3",
    "pygments>=2.13.0",
    "aiohttp>=3.8.1",
    "python-json-logger>=2.0.4",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
iac = "bbrain.iac.cli:cli"

[project.optional-dependencies]
[tool.pdm]
[tool.pdm.build]
includes = [
    "bbrain",
]

[tool.pdm.scripts]
start.cmd = "echo oui"

[tool.pdm.dev-dependencies]
dev = [
    "types-click>=7.1.8",
    "types-requests>=2.28.10",
]
[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"
